+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.1 - 2017-05-19
+
+* Fix incorrect subscription name being created - should have channel prefixed
+
 ## 2.0.0 - 2017-05-17
 
 * Bump up @superbalist/js-pubsub to ^2.0.0